+++ src/btnext.cpp
@@ -311,7 +311,7 @@ CLangBarItemShapeButton::InitMenu (
 		for (i = 0 ; i < ARRAYSIZE (c_rgMenuItems) ; i ++) {
 			wstrDesc		= c_rgMenuItems [i].pchDesc ;
 			if (wstrDesc != NULL) {
-				nstrDesc	= wcslen (wstrDesc) ;
+				nstrDesc	= (ULONG) wcslen (wstrDesc) ;
 				dwFlag		= (i == nCMode)? TF_LBMENUF_CHECKED : 0 ;
 			} else {
 				nstrDesc	= 0 ;
